refactor: SidebarListView
